+++ src/components/Textbox/Textbox.module.css
@@ -68,13 +68,4 @@
     box-shadow: 28rem 0 0 0 var(--tertiary-container);
 }
 
-.time {
-    position: absolute;
-    top: 40%;
-    right: 3rem;
-    color: var(--on-tertiary-container);
-    font-size: 10rem;
-    font-weight: bold;
-
-}
